Make MultiTaskData default data passing mechanism
Moved to data object that can contain more than one dataset as the default data object. The object has a copy method, which is used so that the data object is copied between tasks (not just passing the reference). The default is to copy only the selected dataset, within the multitaskdata object. Setting preserve to True will copy all datasets.